Fix that mods except for `core` cannot be enabled in `mods.json`
Implement `ModIndex::traverse()`, a function which traverses `mod`
folder to construct mod index.
